# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API epesantren pegawai",
        "description": "Dokumentasi API epesantren pegawai",
        "contact": {
            "email": "fp3175723@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/lupa-password": {
            "post": {
                "tags": [
                    "forgot password"
                ],
                "summary": "Forgot Password",
                "description": "Lupa password masukkan kode sekolah dan nip ",
                "operationId": "8ded6f9ab27e4676ff28dc673a745d8c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_sekolah",
                                    "employee_nip"
                                ],
                                "properties": {
                                    "kode_sekolah": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2012001"
                                    },
                                    "employee_nip": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful OTP dikirim di wa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL2xvZ2luIiwiaWF0IjoxNzI5NTY4MTIzLCJleHAiOjE3Mjk1NzE3MjMsIm5iZiI6MTcyOTU2ODEyMywianRpIjoiR1VDR2Z1Y3pPaUtzWlp5QiIsInN1YiI6IjExOCIsInBydiI6IjMyOTYzYTYwNmMyZjE3MWYxYzE0MzMxZTc2OTc2NmNkNTkxMmVkMTUifQ.NuBcDlpsxa8x9XJdsCWeteRmh3Neqa3QlD8ueBmbfec"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data sekolah not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/verifikasi": {
            "post": {
                "tags": [
                    "forgot password"
                ],
                "summary": "Verifikasi OTP",
                "description": "Masukkan kode OTP ",
                "operationId": "ce65efd5827288d3c7a3b72e3f0cc4d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "otp"
                                ],
                                "properties": {
                                    "otp": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2012001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful OTP dikirim di wa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OTP tidak ditemukan"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/update-password": {
            "post": {
                "tags": [
                    "forgot password"
                ],
                "summary": "update Password",
                "description": "Update password baru ",
                "operationId": "636353070280924afcc452a9073b0e9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_password",
                                    "employee_password_confirmed"
                                ],
                                "properties": {
                                    "employee_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "employee_password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password berhasil diganti silahkan login kembali",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "password berhasil diganti silahkan login kembali"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "token has expired"
                    },
                    "400": {
                        "description": "The employee password field corfirmation does not match"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}